Refactor: TickerPrice 관련 API 분리
